+++ lib/invoice_item.rb
@@ -3,8 +3,10 @@ require 'time'
 require 'bigdecimal'
 require_relative './item'
 require_relative './item_repository'
+require_relative 'sanitize'
 
 class InvoiceItem
+include Sanitize
   attr_reader :id,
               :item_id,
               :invoice_id,
@@ -21,20 +23,6 @@ class InvoiceItem
     @unit_price = to_price(info[:unit_price])
     @created_at = to_time(info[:created_at])
     @updated_at = to_time(info[:updated_at])
-  end
-
-  def to_price(price)
-    p = BigDecimal(price, 4)
-    p = (p / 100) if (p % 1).zero?
-    p
-  end
-
-  def to_time(time)
-    Time.parse(time.to_s)
-  end
-
-  def unit_price_to_dollars
-    unit_price.to_f
   end
 
   def update(attributes)
