+++ lib/item.rb
@@ -3,8 +3,10 @@ require 'csv'
 require 'time'
 require 'pry'
 require_relative 'item_repository'
+require_relative 'sanitize'
 
 class Item
+include Sanitize
   attr_reader :id,
               :name,
               :description,
@@ -21,20 +23,6 @@ class Item
     @created_at = to_time(info[:created_at])
     @updated_at = to_time(info[:updated_at])
     @merchant_id = info[:merchant_id].to_i
-  end
-
-  def to_price(price)
-    p = BigDecimal(price, 4)
-    p = (p / 100) if (p % 1).zero?
-    p
-  end
-
-  def to_time(time)
-    Time.parse(time.to_s)
-  end
-
-  def unit_price_to_dollars
-    unit_price.to_f
   end
 
   def update(attributes)
